Moved openFPGA-cores-inventory models into own directory and namespace
- renamed Repo to Repository
- renamed Image Pack to Platform Image Pack
