+++ src/models/OpenFPGA_Cores_Inventory/Platform.cs
@@ -1,4 +1,4 @@
-namespace Pannella.Models;
+namespace Pannella.Models.OpenFPGA_Cores_Inventory;
 
 public class Platform
 {
