+++ src/models/OpenFPGA_Cores_Inventory/Repository.cs
@@ -1,6 +1,6 @@
-namespace Pannella.Models;
+namespace Pannella.Models.OpenFPGA_Cores_Inventory;
 
-public class Repo
+public class Repository
 {
     public string name { get; set; }
     public string owner { get; set; }
