+++ src/models/OpenFPGA_Cores_Inventory/Sponsor.cs
@@ -1,6 +1,6 @@
 using System.Text;
 
-namespace Pannella.Models;
+namespace Pannella.Models.OpenFPGA_Cores_Inventory;
 
 public class Sponsor
 {
