+++ src/partials/Program.Menus.cs
@@ -1,7 +1,7 @@
 using ConsoleTools;
 using Pannella.Helpers;
-using Pannella.Models;
 using Pannella.Models.Extras;
+using Pannella.Models.OpenFPGA_Cores_Inventory;
 using Pannella.Models.Settings;
 using Pannella.Services;
