+++ src/partials/Program.MissingCores.cs
@@ -1,5 +1,5 @@
 using Pannella.Helpers;
-using Pannella.Models;
+using Pannella.Models.OpenFPGA_Cores_Inventory;
 
 namespace Pannella;
