+++ src/services/PocketExtrasService.cs
@@ -5,6 +5,7 @@ using Pannella.Helpers;
 using Pannella.Models;
 using Pannella.Models.Extras;
 using Pannella.Models.Github;
+using Pannella.Models.OpenFPGA_Cores_Inventory;
 using File = System.IO.File;
 
 namespace Pannella.Services;
